% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cengenAnnotate-package.R
\docType{package}
\name{cengenAnnotate-package}
\alias{cengenAnnotate}
\alias{cengenAnnotate-package}
\title{cengenAnnotate: Automated CeNGEN-Based Annotation of Seurat Clusters}
\description{
Automates annotation of C. elegans single-cell RNA-seq clusters by scoring Seurat cluster marker genes against CeNGEN reference expression data. Computes a dot-plot-equivalent coherence score for each candidate cell type from unthresholded expression and percent-expressing values, and classifies clusters as confidently annotatable or in need of manual review, replacing the manual workflow of pasting marker genes into the CeNGEN website and visually reading the resulting dot plot. Reference datasets are discovered and fetched at runtime from a 'pins' board so new CeNGEN dataset releases (developmental stage and sex variants) become available without package changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cengenproject/cengenAnnotate}
  \item Report bugs at \url{https://github.com/cengenproject/cengenAnnotate/issues}
}

}
\author{
\strong{Maintainer}: Marc Hammarlund \email{marc.hammarlund@yale.edu}

}
\keyword{internal}
